refactoring first function chance

+++ src/test/java/YatzyTest.java
@@ -6,9 +6,9 @@ public class YatzyTest {
     @Test
     public void chance_scores_sum_of_all_dice() {
         int expected = 15;
-        int actual = Yatzy.chance(2,3,4,5,1);
+        int actual = Yatzy.chance((byte)2, (byte)3, (byte)4, (byte)5, (byte)1);
         assertEquals(expected, actual);
-        assertEquals(16, Yatzy.chance(3,3,4,5,1));
+        assertEquals(16, Yatzy.chance((byte) 3, (byte)3, (byte)4, (byte)5, (byte)1));
     }
 
     @Test public void yatzy_scores_50() {
